Rename components/manage-flows to components/flows-new
Replace components/flows with components/flows-new after update of flows UI is complete.
