+++ quick-start/src/main/ui/app/app.routes.ts
@@ -4,16 +4,16 @@ import { FlowsComponent } from './components/flows';
 import { EntityModelerComponent } from './components/entity-modeler/entity-modeler.component';
 import { MapComponent } from './components/mappings';
 import { MappingsComponent } from './components/mappings';
-import { ManageFlowsComponent } from './components/manage-flows/manage-flows.component';
+import { ManageFlowsComponent } from './components/flows-new/manage-flows.component';
 import { LoginComponent } from './components/login';
 import { JobsComponent } from './components/jobs';
 import { SettingsComponent } from './components/settings';
 import { TracesComponent, TraceViewerComponent } from './components/traces';
 import { SearchComponent, SearchViewerComponent } from './components/search';
 import { NoContentComponent } from './components/no-content';
-import { EditFlowComponent } from './components/manage-flows/edit-flow/edit-flow.component';
+import { EditFlowComponent } from './components/flows-new/edit-flow/edit-flow.component';
 import { AuthGuard } from './services/auth/auth-guard.service';
-import {FlowsPageComponent} from "./components/manage-flows/manage-flows/flows-page.component";
+import {FlowsPageComponent} from "./components/flows-new/manage-flows/flows-page.component";
 
 export const ROUTES: Routes = [
   { path: '', component: DashboardComponent, canActivate: [AuthGuard] },
